+++ lib/authrocket.rb
@@ -4,7 +4,7 @@ require 'ncore'
   require "authrocket/api/#{f}"
 end
 
-%w(app_hook credential event login_policy membership org realm user user_token).each do |f|
+%w(app_hook auth_provider credential event login_policy membership org realm user user_token).each do |f|
   require "authrocket/#{f}"
 end