+++ lib/authrocket/auth_provider.rb
@@ -1,0 +1,11 @@
+module AuthRocket
+  class AuthProvider < Resource
+    crud :all, :find, :create, :update, :delete
+
+    belongs_to :realm
+
+    attr :provider_type
+    attr :login, :name_field, :signup
+
+  end
+end