+++ lib/authrocket/login_policy.rb
@@ -5,9 +5,10 @@ module AuthRocket
     belongs_to :realm
     has_many :events
 
-    attr :custom_domains, :enable_logins, :enable_signups, :external_css
-    attr :footer, :header, :login_handler, :name, :name_field, :primary_domain
+    attr :custom_domains, :external_css
+    attr :footer, :header, :login_handler, :name, :primary_domain
     attr :signup_handler, :subdomain
+    attr :enable_logins, :enable_signups, :name_field # deprecated
     attr :base_domain, :domains # readonly
 
   end